+++ src/controllers/SessionController.js
@@ -3,6 +3,11 @@ const {Op} = require('sequelize')
 
 module.exports = {
     async read(request, response){
+        const {key} = request.params
+        if(key !== 'ZmluYW1hc3Nh'){
+            response.status(403).send('Acesso restrito!')
+            return
+        }
         const res = await Session.findAll({where:{ativa:1}})
         if(res.length > 0){
             return response.status(200).json(true)
@@ -11,7 +16,11 @@ module.exports = {
     },
 
     async insert(request, response){
-        const {situacao} = request.params
+        const {situacao, key} = request.params
+        if(key !== 'ZmluYW1hc3Nh'){
+            response.status(403).send('Acesso restrito!')
+            return
+        }
         const hasExist = await Session.findAll({where:{ativa:{[Op.eq]:situacao}}})
         if(hasExist > 0){
             return response.status(200).json(true)
@@ -25,7 +34,11 @@ module.exports = {
     },
 
     async update(request, response){
-        const {situacao} = request.params
+        const {situacao, key} = request.params
+        if(key !== 'ZmluYW1hc3Nh'){
+            response.status(403).send('Acesso restrito!')
+            return
+        }
         const res = await Session.update({ativa:situacao},{where:{id:{[Op.gt]:0}}})
         if(res > 0){
             return response.status(200).json(true)
@@ -34,6 +47,11 @@ module.exports = {
     },
 
     async destroy(request, response){
+        const {key} = request.params
+        if(key !== 'ZmluYW1hc3Nh'){
+            response.status(403).send('Acesso restrito!')
+            return
+        }
         const res = await Session.destroy({where:{id:{[Op.gt]:0}}})
         if(res > 0){
             return response.status(200).json(true)
